+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.17.16
 
 * deprecated `call_and_return_stdout`
+* deprecated `retry(call_on_failure=)`